rename navbar component
